# apps/api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DeployCore Control Plane API",
    "version": "0.32.0",
    "description": "Public/control-plane HTTP API for DeployCore (`apps/api`).\nBase path: `/api/v1`. Spec: `GET /openapi.json`."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
...
  "paths": {
...
    "/api/v1/agents/logs": {
      "post": {
        "operationId": "postAgentsLogs",
        "tags": [
          "Agents"
        ],
        "summary": "POST /agents/logs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JSONObject"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "AgentBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JSONObject"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/applications/{applicationId}/logs": {
      "get": {
        "operationId": "getApplicationsByApplicationidLogs",
        "tags": [
          "Applications"
        ],
        "summary": "GET /applications/{applicationId}/logs",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JSONObject"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/deployments/{deploymentId}/events/stream": {
      "get": {
        "operationId": "getDeploymentsByDeploymentidEventsStream",
        "tags": [
          "Deployments"
        ],
        "summary": "GET /deployments/{deploymentId}/events/stream",
        "parameters": [
          {
            "name": "deploymentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JSONObject"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/deployments/{deploymentId}/logs": {
      "get": {
        "operationId": "getDeploymentsByDeploymentidLogs",
        "tags": [
          "Deployments"
        ],
        "summary": "GET /deployments/{deploymentId}/logs",
        "parameters": [
          {
            "name": "deploymentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JSONObject"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "User access token from /auth/login or /auth/register."
      },
      "AgentBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Agent credential from POST /agents/register."
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/APIError"
          }
        }
      },
      "APIError": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "VALIDATION_ERROR",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "RESOURCE_NOT_FOUND",
          "CONFLICT",
          "INTERNAL_ERROR",
          "SERVICE_UNAVAILABLE",
          "RATE_LIMITED",
          "ORGANIZATION_NOT_FOUND",
          "PROJECT_NOT_FOUND",
          "ENVIRONMENT_NOT_FOUND",
          "SERVER_NOT_FOUND",
          "SERVER_OFFLINE",
          "AGENT_UNAVAILABLE",
          "APPLICATION_NOT_FOUND",
          "DEPLOYMENT_NOT_FOUND",
          "REVISION_NOT_FOUND",
          "REVISION_NOT_READY",
          "DOMAIN_ALREADY_ASSIGNED",
          "SECRET_NOT_FOUND",
          "INSUFFICIENT_RESOURCES",
          "BUILD_TIMEOUT",
          "HEALTH_CHECK_FAILED",
          "DATABASE_NOT_FOUND",
          "VOLUME_NOT_FOUND",
          "VOLUME_IN_USE"
        ]
      },
      "JSONObject": {
        "type": "object",
        "additionalProperties": true
      },
...
    }
  }
}